Fold Demo2 module into Demo.
